requirements.txt
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.6
idna==2.8
praw==6.3.1
prawcore==1.0.1
PyGithub==1.43.8
PyJWT==1.7.1
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.2
